+++ allennlp/models/semantic_parsing/atis/atis_semantic_parser.py
@@ -2,16 +2,14 @@ import logging
 from typing import Any, Dict, List, Tuple
 
 import difflib
-import sqlite3
-from multiprocessing import Process
 import sqlparse
 from overrides import overrides
 import torch
 
-from allennlp.common.file_utils import cached_path
 from allennlp.common.util import pad_sequence_to_length
 from allennlp.data import Vocabulary
 from allennlp.data.fields.production_rule_field import ProductionRuleArray
+from allennlp.semparse.executors import SqlExecutor
 from allennlp.models.model import Model
 from allennlp.modules import Attention, Seq2SeqEncoder, TextFieldEmbedder, \
         Embedding
@@ -90,11 +88,7 @@ class AtisSemanticParser(Model):
         self._action_similarity = Average()
         self._denotation_accuracy = Average()
 
-        # Initialize a cursor to our sqlite database, so we can execute SQL queries for denotation accuracy.
-        self._database_file = cached_path(database_file)
-        self._connection = sqlite3.connect(self._database_file)
-        self._cursor = self._connection.cursor()
-
+        self._executor = SqlExecutor(database_file)
         self._action_padding_index = -1  # the padding value used by IndexField
         num_actions = vocab.get_vocab_size(self._rule_namespace)
         if self._add_action_bias:
@@ -131,7 +125,7 @@ class AtisSemanticParser(Model):
                 actions: List[List[ProductionRuleArray]],
                 linking_scores: torch.Tensor,
                 target_action_sequence: torch.LongTensor = None,
-                sql_queries: List[str] = None) -> Dict[str, torch.Tensor]:
+                sql_queries: List[List[str]] = None) -> Dict[str, torch.Tensor]:
         # pylint: disable=arguments-differ
         """
         We set up the initial state for the decoder, and pass that state off to either a DecoderTrainer,
@@ -158,7 +152,7 @@ class AtisSemanticParser(Model):
             The action sequence for the correct action sequence, where each action is an index into the list
             of possible actions.  This tensor has shape ``(batch_size, sequence_length, 1)``. We remove the
             trailing dimension.
-        sql_queries : List[str], optional (default=None)
+        sql_queries : List[List[str]], optional (default=None)
             A list of the SQL queries that are given during training or validation.
         """
         initial_state = self._get_initial_state(utterance, world, actions, linking_scores)
@@ -196,7 +190,7 @@ class AtisSemanticParser(Model):
             best_final_states = self._beam_search.search(num_steps,
                                                          initial_state,
                                                          self._transition_function,
-                                                         keep_final_unfinished_states=False)
+                                                         keep_final_unfinished_states=True)
             outputs['best_action_sequence'] = []
             outputs['debug_info'] = []
             outputs['entities'] = []
@@ -233,34 +227,17 @@ class AtisSemanticParser(Model):
                     self._action_similarity(similarity.ratio())
 
                 if sql_queries and sql_queries[i]:
-                    # Since the query might hang, we run in another process and kill it if it
-                    # takes too long.
-                    process = Process(target=self._sql_result_match,
-                                      args=(predicted_sql_query, sql_queries[i]))
-                    process.start()
-
-                    # If the query has not finished in 3 seconds then we will proceed.
-                    process.join(3)
-                    denotation_correct = process.exitcode # type: ignore
-
-                    if process.is_alive():
-                        logger.info("Evaluating query took over 3 seconds, skipping query")
-                        process.terminate()
-                        process.join()
-
-                    if denotation_correct is None:
-                        denotation_correct = 0
-
+                    denotation_correct = self._executor.evaluate_sql_query(predicted_sql_query, sql_queries[i])
                     self._denotation_accuracy(denotation_correct)
                     outputs['sql_queries'].append(sql_queries[i])
 
-                    outputs['utterance'].append(world[i].utterances[-1])
-                    outputs['tokenized_utterance'].append([token.text
-                                                           for token in world[i].tokenized_utterances[-1]])
-                    outputs['entities'].append(world[i].entities)
-                    outputs['best_action_sequence'].append(action_strings)
-                    outputs['predicted_sql_query'].append(sqlparse.format(predicted_sql_query, reindent=True))
-                    outputs['debug_info'].append(best_final_states[i][0].debug_info[0])  # type: ignore
+                outputs['utterance'].append(world[i].utterances[-1])
+                outputs['tokenized_utterance'].append([token.text
+                                                       for token in world[i].tokenized_utterances[-1]])
+                outputs['entities'].append(world[i].entities)
+                outputs['best_action_sequence'].append(action_strings)
+                outputs['predicted_sql_query'].append(sqlparse.format(predicted_sql_query, reindent=True))
+                outputs['debug_info'].append(best_final_states[i][0].debug_info[0])  # type: ignore
             return outputs
 
     def _get_initial_state(self,
@@ -384,51 +361,11 @@ class AtisSemanticParser(Model):
         return predicted_tensor.equal(targets_trimmed)
 
     @staticmethod
-    def _postprocess_query_sqlite(query: str):
-        # The dialect of SQL that SQLite takes is not exactly the same as the labeled data.
-        # We strip off the parentheses that surround the entire query here.
-        # TODO(kevin): also move this to the SQL executor.
-        query = query.strip()
-        if query.startswith('('):
-            return query[1:query.rfind(')')] + ';'
-        return query
-
-
-    @staticmethod
     def is_nonterminal(token: str):
         if token[0] == '"' and token[-1] == '"':
             return False
         return True
 
-    def _sql_result_match(self, predicted_query: str, sql_query_labels: List[str]) -> int:
-        """
-        We evaluate here whether the predicted query and the query label evaluate to the
-        exact same table. This method is only called by the subprocess, so we just exit with
-        1 if it is correct and 0 otherwise.
-        """
-        # TODO(kevin): move the SQL execution to an executor.
-        postprocessed_predicted_query = self._postprocess_query_sqlite(predicted_query)
-
-        try:
-            self._cursor.execute(postprocessed_predicted_query)
-            predicted_rows = self._cursor.fetchall()
-        except sqlite3.Error as error:
-            logger.debug(f'Error executing predicted: {error}')
-            exit(0)
-
-        # If predicted table matches any of the reference tables then it is counted as correct.
-        target_rows = None
-        for sql_query_label in sql_query_labels:
-            postprocessed_sql_query_label = self._postprocess_query_sqlite(sql_query_label)
-            try:
-                self._cursor.execute(postprocessed_sql_query_label)
-                target_rows = self._cursor.fetchall()
-            except sqlite3.Error as error:
-                logger.debug(f'Error executing predicted: {error}')
-
-            if predicted_rows == target_rows:
-                exit(1)
-        exit(0)
 
     @overrides
     def get_metrics(self, reset: bool = False) -> Dict[str, float]:
