+++ allennlp/semparse/executors/__init__.py
@@ -1,1 +1,2 @@
 from allennlp.semparse.executors.wikitables_sempre_executor import WikiTablesSempreExecutor
+from allennlp.semparse.executors.sql_executor import SqlExecutor
